enhance: ui/ux navigation overhaul mcp servers/registries

+++ ui/user/src/lib/components/mcp/RegistriesView.svelte
@@ -10,18 +10,25 @@
 		getAdminMcpServerAndEntries
 	} from '$lib/context/admin/mcpServerAndEntries.svelte';
 	import {
+		fetchUserMcpServerAndEntries,
+		getUserMcpServerAndEntries
+	} from '$lib/context/mcpServerAndEntries.svelte';
+	import {
 		AdminService,
 		ChatService,
 		type MCPCatalog,
 		type MCPCatalogEntry,
 		type MCPCatalogServer,
 		type OrgUser,
-		MCPCompositeDeletionDependencyError
+		MCPCompositeDeletionDependencyError,
+		Group,
+		type MCPServerInstance
 	} from '$lib/services';
 	import {
 		convertEntriesAndServersToTableData,
 		getServerTypeLabelByType
 	} from '$lib/services/chat/mcp';
+	import { profile } from '$lib/stores';
 	import { formatTimeAgo } from '$lib/time';
 	import { setSearchParamsToLocalStorage } from '$lib/url';
 	import { openUrl } from '$lib/utils';
@@ -31,18 +38,23 @@
 		CircleFadingArrowUp,
 		Ellipsis,
 		LoaderCircle,
+		MessageCircle,
+		SatelliteDish,
 		Server,
 		Trash2
 	} from 'lucide-svelte';
 	import type { Snippet } from 'svelte';
 	import { slide } from 'svelte/transition';
+	import ConnectToServer from '$lib/components/mcp/ConnectToServer.svelte';
 
 	type Item = ReturnType<typeof convertEntriesAndServersToTableData>[number];
 
 	interface Props {
+		entity?: 'workspace' | 'catalog';
+		id?: string;
 		catalog?: MCPCatalog;
 		readonly?: boolean;
-		emptyContentButton?: Snippet;
+		noDataContent?: Snippet;
 		usersMap?: Map<string, OrgUser>;
 		query?: string;
 		urlFilters?: Record<string, (string | number)[]>;
@@ -50,19 +62,24 @@
 		onClearAllFilters?: () => void;
 		onSort?: InitSortFn;
 		initSort?: InitSort;
+		classes?: {
+			tableHeader?: string;
+		};
 	}
 
 	let {
+		entity = 'catalog',
 		catalog = $bindable(),
 		readonly,
-		emptyContentButton,
+		noDataContent,
 		usersMap,
 		query,
 		urlFilters: filters,
 		onFilter,
 		onClearAllFilters,
 		onSort,
-		initSort
+		initSort,
+		classes
 	}: Props = $props();
 
 	let deletingEntry = $state<MCPCatalogEntry>();
@@ -72,14 +89,26 @@
 	let loadingBulkDelete = $state(false);
 	let deleteConflictError = $state<MCPCompositeDeletionDependencyError | undefined>();
 
-	const mcpServerAndEntries = getAdminMcpServerAndEntries();
+	let connectToServerDialog = $state<ReturnType<typeof ConnectToServer>>();
+
+	const mcpServerAndEntries =
+		entity === 'workspace' ? getUserMcpServerAndEntries() : getAdminMcpServerAndEntries();
 	let tableData = $derived(
 		convertEntriesAndServersToTableData(
 			mcpServerAndEntries.entries,
 			mcpServerAndEntries.servers,
-			usersMap
+			usersMap,
+			mcpServerAndEntries.userConfiguredServers
 		)
 	);
+
+	let instances = $derived(mcpServerAndEntries.userServerInstances);
+	let instancesMap = $derived(
+		new Map(instances.map((instance) => [instance.mcpServerID, instance]))
+	);
+
+	let entriesMap = $derived(new Map(mcpServerAndEntries.entries.map((entry) => [entry.id, entry])));
+
 	let filteredTableData = $derived.by(() => {
 		const sorted = tableData.sort((a, b) => {
 			return a.name.localeCompare(b.name);
@@ -94,16 +123,39 @@
 	});
 
 	function getAuditLogsUrl(d: Item) {
-		const isCatalogEntry = d.type === 'single' || d.type === 'remote' || d.type === 'composite';
-		if (isCatalogEntry) {
-			return d.data.powerUserWorkspaceID
-				? `/admin/mcp-servers/w/${d.data.powerUserWorkspaceID}/c/${d.id}?view=audit-logs`
-				: `/admin/mcp-servers/c/${d.id}?view=audit-logs`;
+		let useAdminUrl =
+			window.location.pathname.includes('/admin') && profile.current.hasAdminAccess?.();
+		let hasAuditLogUrlsAccess = profile.current.groups.includes(Group.POWERUSER);
+
+		if (!hasAuditLogUrlsAccess) {
+			return null;
 		}
 
-		return d.data.powerUserWorkspaceID
-			? `/admin/mcp-servers/w/${d.data.powerUserWorkspaceID}/s/${d.id}?view=audit-logs`
-			: `/admin/mcp-servers/s/${d.id}?view=audit-logs`;
+		const isCatalogEntry = d.type === 'single' || d.type === 'remote' || d.type === 'composite';
+		if (isCatalogEntry) {
+			if (useAdminUrl) {
+				return d.data.powerUserWorkspaceID
+					? `/admin/mcp-servers/w/${d.data.powerUserWorkspaceID}/c/${d.id}?view=audit-logs`
+					: `/admin/mcp-servers/c/${d.id}?view=audit-logs`;
+			}
+
+			return `/mcp-servers/c/${d.id}?view=audit-logs`;
+		}
+
+		if (useAdminUrl) {
+			return d.data.powerUserWorkspaceID
+				? `/admin/mcp-servers/w/${d.data.powerUserWorkspaceID}/s/${d.id}?view=audit-logs`
+				: `/admin/mcp-servers/s/${d.id}?view=audit-logs`;
+		}
+		return `/mcp-servers/s/${d.id}?view=audit-logs`;
+	}
+
+	async function fetch() {
+		if (catalog) {
+			await fetchMcpServerAndEntries(catalog.id, mcpServerAndEntries);
+		} else {
+			await fetchUserMcpServerAndEntries(mcpServerAndEntries);
+		}
 	}
 </script>
 
@@ -113,22 +165,12 @@
 			<LoaderCircle class="size-6 animate-spin" />
 		</div>
 	{:else if mcpServerAndEntries.entries.length + mcpServerAndEntries.servers.length === 0}
-		<div class="my-12 flex w-md flex-col items-center gap-4 self-center text-center">
-			<Server class="text-surface3 size-24" />
-			<h4 class="text-on-surface1 text-lg font-semibold">No created MCP servers</h4>
-			<p class="text-on-surface1 text-sm font-light">
-				Looks like you don't have any servers created yet. <br />
-				Click the button below to get started.
-			</p>
-
-			{#if !readonly && emptyContentButton}
-				{@render emptyContentButton()}
-			{/if}
-		</div>
+		{#if noDataContent}
+			{@render noDataContent?.()}
+		{/if}
 	{:else}
-		{@const hasErrors = Object.keys(catalog?.syncErrors ?? {})}
-
-		{#if hasErrors.length && !catalog?.isSyncing}
+		{@const hasCatalogErrors = catalog && Object.keys(catalog?.syncErrors ?? {}).length > 0}
+		{#if hasCatalogErrors && !catalog?.isSyncing}
 			<div class="w-full p-4" in:slide={{ axis: 'y' }} out:slide={{ axis: 'y', duration: 0 }}>
 				<div class="notification-alert flex w-full items-center gap-2 rounded-md p-3 text-sm">
 					<AlertTriangle class="size-" />
@@ -144,14 +186,25 @@
 			{filters}
 			onClickRow={(d, isCtrlClick) => {
 				let url = '';
-				if (d.type === 'single' || d.type === 'remote' || d.type === 'composite') {
-					url = d.data.powerUserWorkspaceID
-						? `/admin/mcp-servers/w/${d.data.powerUserWorkspaceID}/c/${d.id}`
-						: `/admin/mcp-servers/c/${d.id}`;
+				let useAdminUrl =
+					window.location.pathname.includes('/admin') && profile.current.hasAdminAccess?.();
+
+				const id = 'catalogEntryID' in d.data ? d.data.catalogEntryID : d.id;
+				if (useAdminUrl) {
+					if (d.type === 'single' || d.type === 'remote' || d.type === 'composite') {
+						url = d.data.powerUserWorkspaceID
+							? `/admin/mcp-servers/w/${d.data.powerUserWorkspaceID}/c/${id}`
+							: `/admin/mcp-servers/c/${id}`;
+					} else {
+						url = d.data.powerUserWorkspaceID
+							? `/admin/mcp-servers/w/${d.data.powerUserWorkspaceID}/s/${id}`
+							: `/admin/mcp-servers/s/${id}`;
+					}
 				} else {
-					url = d.data.powerUserWorkspaceID
-						? `/admin/mcp-servers/w/${d.data.powerUserWorkspaceID}/s/${d.id}`
-						: `/admin/mcp-servers/s/${d.id}`;
+					url =
+						d.type === 'single' || d.type === 'remote' || d.type === 'composite'
+							? `/mcp-servers/c/${id}`
+							: `/mcp-servers/s/${id}`;
 				}
 
 				setSearchParamsToLocalStorage(page.url.pathname, page.url.search);
@@ -165,7 +218,7 @@
 			noDataMessage="No catalog servers added."
 			classes={{
 				root: 'rounded-none rounded-b-md shadow-none',
-				thead: 'top-31'
+				thead: classes?.tableHeader
 			}}
 			validateSelect={(d) => d.editable}
 			disabledSelectMessage="This entry is managed by Git; changes cannot be made."
@@ -204,40 +257,88 @@
 				{/if}
 			{/snippet}
 			{#snippet actions(d)}
-				{@const url = getAuditLogsUrl(d)}
+				{@const auditLogUrl = getAuditLogsUrl(d)}
+				{@const canDelete = d.editable && !readonly}
 				<DotDotDot class="icon-button hover:dark:bg-background/50">
 					{#snippet icon()}
 						<Ellipsis class="size-4" />
 					{/snippet}
 
-					<div class="default-dialog flex min-w-max flex-col gap-1 p-2">
-						<button
-							onclick={(e) => {
-								e.stopPropagation();
-								const isCtrlClick = e.ctrlKey || e.metaKey;
-								setSearchParamsToLocalStorage(page.url.pathname, page.url.search);
-								openUrl(url, isCtrlClick);
-							}}
-							class="menu-button"
-						>
-							<Captions class="size-4" /> View Audit Logs
-						</button>
-						{#if d.editable && !readonly}
+					{#snippet children({ toggle })}
+						<div class="default-dialog flex min-w-max flex-col gap-1 p-2">
 							<button
-								class="menu-button-destructive"
-								onclick={(e) => {
+								class="menu-button-primary"
+								onclick={async (e) => {
 									e.stopPropagation();
-									if (d.data.type === 'mcpserver') {
-										deletingServer = d.data as MCPCatalogServer;
-									} else {
-										deletingEntry = d.data as MCPCatalogEntry;
-									}
+									const entry =
+										'catalogEntryID' in d.data
+											? entriesMap.get(d.data.catalogEntryID as string)
+											: 'isCatalogEntry' in d.data
+												? d.data
+												: undefined;
+									const server = 'isCatalogEntry' in d.data ? undefined : d.data;
+									const instance =
+										'isCatalogEntry' in d.data
+											? undefined
+											: d.data.catalogEntryID
+												? undefined
+												: instancesMap.get(d.id);
+									connectToServerDialog?.open({
+										entry,
+										server,
+										instance
+									});
+									toggle(false);
 								}}
 							>
-								<Trash2 class="size-4" /> Delete Server
+								<SatelliteDish class="size-4" /> Connect
 							</button>
-						{/if}
-					</div>
+							{#if !('isCatalogEntry' in d.data)}
+								<button
+									class="menu-button"
+									onclick={async (e) => {
+										e.stopPropagation();
+										const serverToUse = 'isCatalogEntry' in d.data ? undefined : d.data;
+										if (serverToUse) {
+											connectToServerDialog?.handleSetupChat(serverToUse, instancesMap.get(d.id));
+										}
+										toggle(false);
+									}}
+								>
+									<MessageCircle class="size-4" /> Chat
+								</button>
+							{/if}
+
+							{#if auditLogUrl}
+								<button
+									onclick={(e) => {
+										e.stopPropagation();
+										const isCtrlClick = e.ctrlKey || e.metaKey;
+										setSearchParamsToLocalStorage(page.url.pathname, page.url.search);
+										openUrl(auditLogUrl, isCtrlClick);
+									}}
+									class="menu-button"
+								>
+									<Captions class="size-4" /> View Audit Logs
+								</button>
+							{/if}
+							{#if canDelete}
+								<button
+									class="menu-button-destructive"
+									onclick={(e) => {
+										e.stopPropagation();
+										if (d.data.type === 'mcpserver') {
+											deletingServer = d.data as MCPCatalogServer;
+										} else {
+											deletingEntry = d.data as MCPCatalogEntry;
+										}
+									}}
+								>
+									<Trash2 class="size-4" /> Delete Server
+								</button>
+							{/if}
+						</div>
+					{/snippet}
 				</DotDotDot>
 			{/snippet}
 			{#snippet tableSelectActions(currentSelected)}
@@ -262,7 +363,7 @@
 	names={[deletingEntry?.manifest?.name ?? '']}
 	show={Boolean(deletingEntry)}
 	onsuccess={async () => {
-		if (!deletingEntry || !catalog) {
+		if (!deletingEntry) {
 			return;
 		}
 
@@ -271,10 +372,11 @@
 				deletingEntry.powerUserWorkspaceID,
 				deletingEntry.id
 			);
-		} else {
+		} else if (catalog) {
 			await AdminService.deleteMCPCatalogEntry(catalog.id, deletingEntry.id);
 		}
-		await fetchMcpServerAndEntries(catalog.id, mcpServerAndEntries);
+
+		await fetch();
 		deletingEntry = undefined;
 	}}
 	oncancel={() => (deletingEntry = undefined)}
@@ -296,10 +398,11 @@
 					deletingServer.powerUserWorkspaceID,
 					deletingServer.id
 				);
-			} else {
+			} else if (catalog) {
 				await AdminService.deleteMCPCatalogServer(catalog.id, deletingServer.id);
 			}
-			await fetchMcpServerAndEntries(catalog.id, mcpServerAndEntries);
+
+			await fetch();
 			deletingServer = undefined;
 		} catch (error) {
 			if (error instanceof MCPCompositeDeletionDependencyError) {
@@ -319,7 +422,6 @@
 	names={Object.values(selected).map((s) => s.name)}
 	show={confirmBulkDelete}
 	onsuccess={async () => {
-		if (!catalog) return;
 		loadingBulkDelete = true;
 		try {
 			for (const item of Object.values(selected)) {
@@ -330,7 +432,7 @@
 								item.data.powerUserWorkspaceID,
 								item.data.id
 							);
-						} else {
+						} else if (catalog) {
 							await AdminService.deleteMCPCatalogServer(catalog.id, item.data.id);
 						}
 					} catch (error) {
@@ -348,12 +450,13 @@
 							item.data.powerUserWorkspaceID,
 							item.data.id
 						);
-					} else {
+					} else if (catalog) {
 						await AdminService.deleteMCPCatalogEntry(catalog.id, item.data.id);
 					}
 				}
 			}
-			await fetchMcpServerAndEntries(catalog.id, mcpServerAndEntries);
+
+			await fetch();
 		} finally {
 			confirmBulkDelete = false;
 			loadingBulkDelete = false;
@@ -371,4 +474,9 @@
 	onClose={() => {
 		deleteConflictError = undefined;
 	}}
+/>
+
+<ConnectToServer
+	bind:this={connectToServerDialog}
+	userConfiguredServers={mcpServerAndEntries.userConfiguredServers}
 />
